Revert "[GPU][Codegen] Expand iteration space based on new `expand_dims` attribute (#22342)"
This reverts commit b879ee18ff93d9c481b2bdaf206d3a20803b4743.
