+++ compiler/src/iree/compiler/Codegen/Common/GPU/Passes.td
@@ -383,14 +383,6 @@ def GPUApplyPaddingLevelPass :
   ];
 }
 
-def GPUExpandDimensionsPass :
-    InterfacePass<"iree-codegen-gpu-expand-dimensions", "mlir::FunctionOpInterface"> {
-  let summary = "Pass to expand tensor op dims based on `expand_dims` lowering_config";
-  let dependentDialects = [
-    "::mlir::iree_compiler::IREE::Util::UtilDialect"
-  ];
-}
-
 def GPUTensorTileToSerialLoopsPass :
     InterfacePass<"iree-codegen-gpu-tensor-tile-to-serial-loops", "mlir::FunctionOpInterface"> {
   let summary = "Pass to tile reduction dimensions for certain GPU ops";
