+++ compiler/src/iree/compiler/Codegen/Common/GPU/test/CMakeLists.txt
@@ -32,7 +32,6 @@ iree_lit_test_suite(
     "gpu_distribute_forall.mlir"
     "gpu_distribute_scf_for.mlir"
     "gpu_distribute_shared_memory.mlir"
-    "gpu_expand_dimensions.mlir"
     "gpu_fuse_and_hoist_forall.mlir"
     "gpu_generalize_named_ops.mlir"
     "gpu_greedily_distribute_to_threads.mlir"
