+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/IR/GPULoweringConfigUtils.h
@@ -62,9 +62,6 @@ IREE::GPU::LoweringConfigAttr setPromotedOperandsList(
 std::optional<SmallVector<int64_t>> getPaddingList(LoweringConfigAttr config,
                                                    bool paddingConv = false);
 
-/// Helper to retrieve dimension expansion config from lowering config.
-DimensionExpansionAttr getDimensionExpansion(LoweringConfigAttr config);
-
 } // namespace mlir::iree_compiler::IREE::GPU
 
 #endif // IREE_COMPILER_CODEGEN_DIALECT_GPU_IR_GPULOWERINGCONFIGUTILS_H_
