+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUAttrs.td
@@ -959,60 +959,4 @@ def IREEGPU_GPUPipelineOptionsAttr : AttrDef<IREEGPU_Dialect, "GPUPipelineOption
   }];
 }
 
-//===----------------------------------------------------------------------===//
-// iree_gpu.expand_dims
-//===----------------------------------------------------------------------===//
-def IREEGPU_DimensionExpansionAttr
-    : AttrDef<IREEGPU_Dialect, "DimensionExpansion"> {
-  let mnemonic = "expand_dims";
-  let cppNamespace = "::mlir::iree_compiler::IREE::GPU";
-
-  let summary = [{Attribute for describing static dimension expansion.}];
-  let description = [{
-    This attribute describes how dimensions in an iteration space should be
-    expanded. Each original dimension can either remain unchanged or be
-    split into multiple dimensions. The semantics are similar to the familiar
-    `tensor.expand_shape` operation.
-
-    The reassociations parameter specifies the mapping from original dimensions
-    to expanded dimensions. For example, [[0], [1], [2, 3]] means:
-    - Original dimension 0 maps to output dimension 0
-    - Original dimension 1 maps to output dimension 1
-    - Original dimension 2 is split into output dimensions 2 and 3
-
-    The output_shape parameter specifies the sizes of the expanded dimensions.
-    If the size is ShapedType::kDynamic, the size is determined from the product
-    of the rest of the static tile sizes in the respective reassociation group.
-    There can be at most one dynamic size per reassociation group.
-
-    Example: #iree_gpu.expand_dims<[[0], [1], [2, 3]], output_shape = [?, ?, ?, 8]>
-  }];
-
-  let parameters = (ins
-    "ArrayAttr":$reassociations,
-    "DenseI64ArrayAttr":$output_shape
-  );
-
-  let builders = [
-    AttrBuilder<(ins
-      "ArrayRef<ReassociationIndices>":$reassociations,
-      "ArrayRef<int64_t>":$outputShape)>
-  ];
-
-  let assemblyFormat = "`<` $reassociations `,` `output_shape` `=` custom<DynamicI64DenseArrayAttr>($output_shape) `>`";
-
-  let extraClassDeclaration = [{
-  SmallVector<ReassociationIndices> getReassociationIndices() {
-  return llvm::to_vector<4>(llvm::map_range(
-      getReassociations().getAsRange<ArrayAttr>(),
-      [](ArrayAttr arrayAttr) -> ReassociationIndices {
-        return llvm::to_vector<2>(llvm::map_range(
-            arrayAttr.getAsRange<IntegerAttr>(),
-            [](IntegerAttr idx) { return idx.getInt(); }));
-      }));
-  }
-  }];
-
-  let genVerifyDecl = 1;
-}
 #endif // IREE_COMPILER_CODEGEN_DIALECT_GPU_IREEGPUATTRS
