+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/TargetUtils/ReductionConfigUtils.cpp
@@ -4,8 +4,6 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Codegen/Dialect/GPU/IR/GPULoweringConfigUtils.h"
-#include "iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUAttrs.h"
 #include "iree/compiler/Codegen/Dialect/GPU/TargetUtils/ConfigUtils.h"
 
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenAttrs.h"
@@ -296,44 +294,10 @@ getVectorDistributeReductionConfig(
   int subgroup = partialReductionSize / subgroupStride;
   int64_t subgroupBasis = (subgroup == 0) ? 1 : subgroup;
 
-  SmallVector<ReassociationIndices> reassociations;
-  SmallVector<int64_t> outputShape;
-
-  // We require the reduction dimension to be evenly divisible by threadLoads
-  // because the current expansion strategy doesn't support padding.
-  if (ShapedType::isStaticShape(bounds) && threadLoads > 1 &&
-      bounds[lastReductionDim] % threadLoads == 0) {
-    workgroupTileSizes.push_back(0);
-    partialReductionTileSizes.push_back(0);
-    threadTileSizes.push_back(0);
-    threadCounts.push_back(1);
-    subGroupCounts.push_back(1);
-    mapping.push_back(mapping.size());
-
-    int64_t outer = lastReductionDim;
-    int64_t inner = lastReductionDim + 1;
-
-    for (int64_t i = 0; i < op.getNumLoops(); ++i) {
-      if (i == lastReductionDim) {
-        int64_t idx = outputShape.size();
-        reassociations.push_back({idx, idx + 1});
-        outputShape.append({ShapedType::kDynamic, threadLoads});
-      } else {
-        reassociations.push_back({static_cast<int64_t>(outputShape.size())});
-        outputShape.push_back(ShapedType::kDynamic);
-      }
-    }
-
-    partialReductionTileSizes[outer] = partialReductionSize / threadLoads;
-    threadTileSizes[inner] = threadLoads;
-    threadCounts[outer] = threadBasis;
-    subGroupCounts[outer] = subgroupBasis;
-  } else {
-    partialReductionTileSizes[lastReductionDim] = partialReductionSize;
-    threadTileSizes[lastReductionDim] = threadLoads;
-    threadCounts[lastReductionDim] = threadBasis;
-    subGroupCounts[lastReductionDim] = subgroupBasis;
-  }
+  partialReductionTileSizes[lastReductionDim] = partialReductionSize;
+  threadTileSizes[lastReductionDim] = threadLoads;
+  threadCounts[lastReductionDim] = threadBasis;
+  subGroupCounts[lastReductionDim] = subgroupBasis;
 
   ArrayAttr subgroupBasisAttr = b.getArrayAttr(
       {b.getI64ArrayAttr(subGroupCounts), b.getI64ArrayAttr(mapping)});
@@ -341,20 +305,13 @@ getVectorDistributeReductionConfig(
   ArrayAttr threadBasisAttr = b.getArrayAttr(
       {b.getI64ArrayAttr(threadCounts), b.getI64ArrayAttr(mapping)});
 
-  SmallVector<NamedAttribute> configAttrs = {
-      b.getNamedAttr("workgroup", b.getI64ArrayAttr(workgroupTileSizes)),
-      b.getNamedAttr("partial_reduction",
+  NamedAttribute configAttrs[] = {
+      NamedAttribute("workgroup", b.getI64ArrayAttr(workgroupTileSizes)),
+      NamedAttribute("partial_reduction",
                      b.getI64ArrayAttr(partialReductionTileSizes)),
-      b.getNamedAttr("thread", b.getI64ArrayAttr(threadTileSizes)),
-      b.getNamedAttr("lane_basis", threadBasisAttr),
-      b.getNamedAttr("subgroup_basis", subgroupBasisAttr),
-  };
-
-  if (!reassociations.empty()) {
-    auto dimExpandAttr =
-        DimensionExpansionAttr::get(context, reassociations, outputShape);
-    configAttrs.emplace_back(b.getNamedAttr("expand_dims", dimExpandAttr));
-  }
+      NamedAttribute("thread", b.getI64ArrayAttr(threadTileSizes)),
+      NamedAttribute("lane_basis", threadBasisAttr),
+      NamedAttribute("subgroup_basis", subgroupBasisAttr)};
 
   auto configDict = b.getDictionaryAttr(configAttrs);
   auto loweringConfig = IREE::GPU::LoweringConfigAttr::get(context, configDict);
