+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.cpp
@@ -736,8 +736,6 @@ void addGPUVectorDistributePassPipeline(OpPassManager &funcPassManager,
   funcPassManager.addPass(createCSEPass());
   funcPassManager.addPass(createGPUPromoteMatmulOperandsPass());
 
-  funcPassManager.addPass(createGPUExpandDimensionsPass());
-
   // Tile to reduction loops.
   {
     GPUApplyTilingLevelPassOptions options;
