+++ compiler/src/iree/compiler/Codegen/LLVMGPU/test/ROCDL/config_vector_distribute_reduction_gfx942.mlir
@@ -84,12 +84,11 @@ func.func @reduction_with_no_consumer() {
 
 // CHECK-LABEL: func.func @reduction_with_no_consumer
 // CHECK:           lowering_config = #iree_gpu.lowering_config
-// CHECK-SAME:      expand_dims = #iree_gpu.expand_dims<{{\[}}[0], [1], [2], [3, 4]{{\]}}, output_shape = [?, ?, ?, ?, 8]>
-// CHECK-SAME:      lane_basis = {{\[}}[1, 1, 1, 64, 1], [0, 1, 2, 3, 4]{{\]}}
-// CHECK-SAME:      partial_reduction = [0, 0, 1, 512, 0]
-// CHECK-SAME:      subgroup_basis = {{\[}}[1, 1, 1, 8, 1], [0, 1, 2, 3, 4]{{\]}}
-// CHECK-SAME:      thread = [0, 0, 1, 1, 8]
-// CHECK-SAME:      workgroup = [1, 1, 0, 0, 0]
+// CHECK-SAME:      lane_basis = {{\[}}[1, 1, 1, 64], [0, 1, 2, 3]
+// CHECK-SAME:      partial_reduction = [0, 0, 1, 4096]
+// CHECK-SAME:      subgroup_basis = {{\[}}[1, 1, 1, 8], [0, 1, 2, 3]
+// CHECK-SAME:      thread = [0, 0, 1, 8],
+// CHECK-SAME:      workgroup = [1, 1, 0, 0]
 
 // -----
 
@@ -151,22 +150,20 @@ func.func @test_multiple_reduction() {
 // CHECK-SAME:    ins(%{{.*}} : tensor<2x32x10x16384xf32>)
 // CHECK-SAME:    outs({{.*}}: tensor<2x32xf32>)
 // CHECK-SAME:    attrs =  {lowering_config = #iree_gpu.lowering_config<{
-// CHECK-SAME:               expand_dims = #iree_gpu.expand_dims<{{\[}}[0], [1], [2], [3, 4]{{\]}}, output_shape = [?, ?, ?, ?, 8]>,
-// CHECK-SAME:               lane_basis = {{\[}}[1, 1, 1, 64, 1], [0, 1, 2, 3, 4]{{\]}},
-// CHECK-SAME:               partial_reduction = [0, 0, 1, 1024, 0],
-// CHECK-SAME:               subgroup_basis = {{\[}}[1, 1, 1, 16, 1], [0, 1, 2, 3, 4]{{\]}},
-// CHECK-SAME:               thread = [0, 0, 1, 1, 8],
-// CHECK-SAME:               workgroup = [1, 1, 0, 0, 0]
+// CHECK-SAME:               lane_basis = {{\[}}[1, 1, 1, 64], [0, 1, 2, 3]],
+// CHECK-SAME:               partial_reduction = [0, 0, 1, 8192],
+// CHECK-SAME:               subgroup_basis = {{\[}}[1, 1, 1, 16], [0, 1, 2, 3]],
+// CHECK-SAME:               thread = [0, 0, 1, 8],
+// CHECK-SAME:               workgroup = [1, 1, 0, 0]
 // CHECK:       %{{.*}} = linalg.generic {indexing_maps = [#map, #map1, #map1],
 // CHECK-SAME:    iterator_types = ["parallel", "parallel", "reduction", "reduction"]}
 // CHECK-SAME:    ins{{.*}}, {{.*}} : tensor<2x32x10x16384xf32>, tensor<2x32xf32>)
 // CHECK-SAME:    outs(%{{.*}} : tensor<2x32xf32>)
 // CHECK-SAME:    attrs =  {lowering_config = #iree_gpu.lowering_config<{
-// CHECK-SAME:              expand_dims = #iree_gpu.expand_dims<{{\[}}[0], [1], [2], [3, 4]{{\]}}, output_shape = [?, ?, ?, ?, 8]>,
-// CHECK-SAME:              lane_basis = {{\[}}[1, 1, 1, 64, 1], [0, 1, 2, 3, 4]{{\]}},
-// CHECK-SAME:              partial_reduction = [0, 0, 1, 1024, 0],
-// CHECK-SAME:              subgroup_basis = {{\[}}[1, 1, 1, 16, 1], [0, 1, 2, 3, 4]{{\]}},
-// CHECK-SAME:              thread = [0, 0, 1, 1, 8],
+// CHECK-SAME:              lane_basis = {{\[}}[1, 1, 1, 64], [0, 1, 2, 3]],
+// CHECK-SAME:              partial_reduction = [0, 0, 1, 8192],
+// CHECK-SAME:              subgroup_basis = {{\[}}[1, 1, 1, 16], [0, 1, 2, 3]],
+// CHECK-SAME:              thread = [0, 0, 1, 8],
 // CHECK:       %{{.*}} = linalg.generic {indexing_maps = [#map, #map1, #map1, #map],
 // CHECK-SAME:    iterator_types = ["parallel", "parallel", "parallel", "parallel"]}
 // CHECK-SAME:    ins({{.*}}, %{{.*}}, {{.*}} : tensor<2x32x10x16384xf16>, tensor<2x32xf32>, tensor<2x32xf32>)
@@ -253,12 +250,11 @@ func.func @test_multiple_stores(%arg0: !iree_tensor_ext.dispatch.tensor<readonly
 //  CHECK-SAME:               workgroup = [1, 0]
 //       CHECK:   linalg.generic
 //  CHECK-SAME:      attrs =  {lowering_config = #iree_gpu.lowering_config<{
-//  CHECK-SAME:               expand_dims = #iree_gpu.expand_dims<{{\[}}[0], [1, 2]{{\]}}, output_shape = [?, ?, 4]>,
-//  CHECK-SAME:               lane_basis = {{\[}}[1, 64, 1], [0, 1, 2]{{\]}},
-//  CHECK-SAME:               partial_reduction = [0, 1024, 0],
-//  CHECK-SAME:               subgroup_basis = {{\[}}[1, 16, 1], [0, 1, 2]{{\]}},
-//  CHECK-SAME:               thread = [0, 1, 4],
-//  CHECK-SAME:               workgroup = [1, 0, 0]
+//  CHECK-SAME:               lane_basis = {{\[}}[1, 64], [0, 1]],
+//  CHECK-SAME:               partial_reduction = [0, 4096],
+//  CHECK-SAME:               subgroup_basis = {{\[}}[1, 16], [0, 1]],
+//  CHECK-SAME:               thread = [0, 4],
+//  CHECK-SAME:               workgroup = [1, 0]
 
 // -----
 
@@ -295,9 +291,9 @@ func.func @test_gather_config(%arg0: !iree_tensor_ext.dispatch.tensor<readonly:t
 //      CHECK:    linalg.yield
 //      CHECK:   linalg.generic
 // CHECK-SAME:      attrs =  {lowering_config = #iree_gpu.lowering_config<{
-// CHECK-SAME:               lane_basis = {{\[}}[1, 64], [0, 1]{{\]}},
+// CHECK-SAME:               lane_basis = {{\[}}[1, 64], [0, 1]],
 // CHECK-SAME:               partial_reduction = [0, 64],
-// CHECK-SAME:               subgroup_basis = {{\[}}[1, 1], [0, 1]{{\]}},
+// CHECK-SAME:               subgroup_basis = {{\[}}[1, 1], [0, 1]],
 // CHECK-SAME:               thread = [0, 1],
 // CHECK-SAME:               workgroup = [1, 0]
 
@@ -436,9 +432,8 @@ func.func @batch_matvec_f16_f32() {
 // CHECK-LABEL: @batch_matvec_f16_f32
 //       CHECK:   linalg.generic
 //  CHECK-SAME:      attrs = {lowering_config = #iree_gpu.lowering_config<{
-//  CHECK-SAME:                 expand_dims = #iree_gpu.expand_dims<{{\[}}[0], [1], [2, 3]{{\]}}, output_shape = [?, ?, ?, 8]>,
-//  CHECK-SAME:                 lane_basis = {{\[}}[1, 1, 64, 1], [0, 1, 2, 3]{{\]}},
-//  CHECK-SAME:                 partial_reduction = [0, 0, 64, 0],
-//  CHECK-SAME:                 subgroup_basis = {{\[}}[1, 1, 1, 1], [0, 1, 2, 3]{{\]}},
-//  CHECK-SAME:                 thread = [0, 0, 1, 8],
-//  CHECK-SAME:                 workgroup = [4, 1, 0, 0]
+//  CHECK-SAME:                 lane_basis = {{\[}}[1, 1, 64], [0, 1, 2]],
+//  CHECK-SAME:                 partial_reduction = [0, 0, 512],
+//  CHECK-SAME:                 subgroup_basis = {{\[}}[1, 1, 1], [0, 1, 2]],
+//  CHECK-SAME:                 thread = [0, 0, 8],
+//  CHECK-SAME:                 workgroup = [4, 1, 0]
