+++ compiler/src/iree/compiler/Dialect/Encoding/IR/BUILD.bazel
@@ -70,7 +70,6 @@ iree_compiler_cc_library(
         ":EncodingTypesGen",
         "//compiler/src/iree/compiler/Dialect/LinalgExt/Utils",
         "//compiler/src/iree/compiler/Dialect/TensorExt/IR",
-        "//compiler/src/iree/compiler/Utils",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:ArithDialect",
         "@llvm-project//mlir:DialectUtils",
