+++ compiler/src/iree/compiler/Dialect/Encoding/IR/CMakeLists.txt
@@ -54,7 +54,6 @@ iree_cc_library(
     MLIRTensorUtils
     iree::compiler::Dialect::LinalgExt::Utils
     iree::compiler::Dialect::TensorExt::IR
-    iree::compiler::Utils
   PUBLIC
 )
