+++ compiler/src/iree/compiler/Utils/BUILD.bazel
@@ -34,7 +34,6 @@ iree_compiler_cc_library(
     name = "Utils",
     srcs = [
         "ConversionUtils.cpp",
-        "EncodingUtils.cpp",
         "EquivalenceUtils.cpp",
         "FlatbufferUtils.cpp",
         "Indexing.cpp",
@@ -50,7 +49,6 @@ iree_compiler_cc_library(
     hdrs = [
         "ConversionUtils.h",
         "EmbeddedDataDirectory.h",
-        "EncodingUtils.h",
         "EquivalenceUtils.h",
         "FlatbufferUtils.h",
         "Folding.h",
