+++ compiler/src/iree/compiler/Utils/CMakeLists.txt
@@ -16,7 +16,6 @@ iree_cc_library(
   HDRS
     "ConversionUtils.h"
     "EmbeddedDataDirectory.h"
-    "EncodingUtils.h"
     "EquivalenceUtils.h"
     "FlatbufferUtils.h"
     "Folding.h"
@@ -35,7 +34,6 @@ iree_cc_library(
     "TracingUtils.h"
   SRCS
     "ConversionUtils.cpp"
-    "EncodingUtils.cpp"
     "EquivalenceUtils.cpp"
     "FlatbufferUtils.cpp"
     "Indexing.cpp"
